Rename test codes.
Renamed to reflect the nature of the test code (for example, UnitTests or IntegrationTests).
