+++ src/test/java/io/vanslog/spring/data/meilisearch/client/ClientConfigurationUnitTest.java
@@ -14,36 +14,31 @@
  * limitations under the License.
  */
 
-package io.vanslog.spring.data.meilisearch.core.convert;
+package io.vanslog.spring.data.meilisearch.client;
 
-import static org.assertj.core.api.Assertions.*;
-
-import io.vanslog.spring.data.meilisearch.core.mapping.SimpleMeilisearchMappingContext;
+import static org.assertj.core.api.AssertionsForClassTypes.*;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 /**
- * Unit tests for {@link MappingMeilisearchConverter}.
+ * Tests for {@link ClientConfiguration}.
  *
  * @author Junghoon Ban
  */
-class MappingMeilisearchConverterTest {
+class ClientConfigurationUnitTest {
 
-	private MappingMeilisearchConverter converter;
+	ClientConfiguration clientConfiguration;
 
 	@BeforeEach
-	void setUp() {
-		converter = new MappingMeilisearchConverter(new SimpleMeilisearchMappingContext());
+	void setup() {
+		clientConfiguration = ClientConfiguration.builder().connectedToLocalhost().withApiKey("masterKey").build();
 	}
 
 	@Test
-	void shouldReturnMappingContext() {
-		assertThat(converter.getMappingContext()).isNotNull();
-	}
-
-	@Test
-	void shouldReturnConversionService() {
-		assertThat(converter.getConversionService()).isNotNull();
+	void shouldBuildConfiguration() {
+		assertThat(clientConfiguration.getHostUrl()).isEqualTo("http://localhost:7700");
+		assertThat(clientConfiguration.getApiKey()).isEqualTo("masterKey");
+		assertThat(clientConfiguration.getClientAgents()).isEmpty();
 	}
 }
