+++ src/test/java/io/vanslog/spring/data/meilisearch/repository/support/MeilisearchEntityInformationCreatorImplUnitTests.java
@@ -32,7 +32,7 @@ import org.springframework.data.mapping.MappingException;
  *
  * @author Junghoon Ban
  */
-class MeilisearchEntityInformationCreatorImplTest {
+class MeilisearchEntityInformationCreatorImplUnitTests {
 
 	MeilisearchEntityInformationCreatorImpl entityInformationCreator;
